разбить код из main по модулям
